Cleanup settings components
